* commit '71e98878396b3bca0af5c8b55f3063354df762f6':
  Small changes based on PR comments and theme based CSS is now generated by the backend
  [ITB-46] Provided customizable legal notices
  Implemented improved authentication and session management and small fixes
  Small fixes in managing sessions and shortcut test case selection
  [ITB-43] Implemented shortcut test case selection for vendor users
  [ITB-41] Introduced commission theme
  [ITB-44] Commented out tutorial navigation menu item
  [ITB-40] Allowed test cases with no end-user configuration
  [ITB-38] Added option on dashboard to set max idle time of sessions
  [ITB-37] Added action button on dashboard to terminate running tests
  [ITB-36] Added action button on dashboard to view test reports
  [ITB-35] Provided dashboard overview of test sessions
  [ITB-9] Implemented scheduling job to kill idle sessions

+++ gitb-engine/src/main/java/com/gitb/engine/SessionManager.java
@@ -31,6 +31,14 @@ public class SessionManager {
 		return instance;
 	}
 
+	public boolean exists(String sessionId) {
+		return contexts.containsKey(sessionId);
+	}
+
+	public boolean notExists(String sessionId) {
+		return !exists(sessionId);
+	}
+
     /**
      * Create a new testcase execution session
      * @param testCaseId
@@ -65,7 +73,9 @@ public class SessionManager {
 	public void endSession(String sessionId) {
 		TestCaseContext testCaseContext = contexts.remove(sessionId);
 
-		testCaseContext.destroy();
+		if (testCaseContext != null) {
+			testCaseContext.destroy();
+		}
 	}
 
 	public void destroy() {
